'deserialize' function with 4/4 tests passed.
Not really happy with this solution as there is a reliance on the order
object properties are defined and iterated.
Even as a string, you could have row0_name, row1_name, row0_value, row1_value...
and identify repeating patterns - but which is the value and which is the
property name? (They will be equal in number) and what if a value is the same as a key?